✨ [feat] 캘린더 지출목록 지출 및 예산 계산 기능 구현
- 캘린더 셀의 기준의 전 지출 값과 예산을 뺀 값을 남을 예산 레이블에 넣어주는 기능 구현
- 소수점 두자리까지 보여주는 포맷팅 구현 (Double로 리팩토링)

+++ TripLog/TripLog/Feat-Jamong/View/CalendarExpenseView/CalendarExpenseView.swift
@@ -98,9 +98,9 @@ final class CalendarExpenseView: UIView {
     ///   - date: 선택된 날짜
     ///   - expenses: 해당 날짜의 지출 항목 배열
     ///   - balance: 현재 잔액
-    func configure(date: Date, expenses: [MockMyCashBookModel], balance: Int) {
+    func configure(date: Date, expenses: [MockMyCashBookModel], balance: Double) {
         self.expenses = expenses
-        let totalExpense = Int(expenses.reduce(0) { $0 + $1.amount })
+        let totalExpense = Double(expenses.reduce(0) { $0 + $1.amount })
         headerView.configure(date: date, expense: totalExpense, balance: balance)
         
         emptyStateLabel.isHidden = !expenses.isEmpty
